Backend replay op generation pipeline to instrumentation library complete

+++ r3/common/src/trace.rs
@@ -5,8 +5,8 @@ use serde::{Serialize, Deserialize};
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone, Copy)]
 pub enum CallID {
     ScUnknown,
-    ScMmap { grow: i32 },
-    ScWritev { fd: i32, iov: i32, iovcnt: i32 },
+    ScMmap { grow: u32 },
+    ScWritev { fd: i32, iov: i32, iovcnt: u32 },
     ScThreadSpawn,
     ScFutex,
     ScExit,
@@ -16,10 +16,10 @@ pub enum CallID {
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
 pub struct Access {
-    pub access_idx: i32,
+    pub access_idx: u32,
     pub opcode: i32,
     pub addr: i32,
-    pub size: i32,
+    pub size: u32,
     pub load_value: i64,
     pub expected_value: i64,
 }
@@ -32,9 +32,9 @@ impl fmt::Display for Access {
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
 pub struct Call {
-    pub access_idx: i32,
+    pub access_idx: u32,
     pub opcode: i32,
-    pub func_idx: i32,
+    pub func_idx: u32,
     pub return_val: i64,
     pub call_id: CallID,
 }
@@ -78,11 +78,11 @@ impl<'a> TraceDataDeser<'a> {
 }
 
 /* Convert instrumentation-level CallID to Rust Enum */
-pub fn create_call_id(call_id: i32, args: [i32;3]) -> Option<CallID> {
+pub fn create_call_id(call_id: u32, args: [i64;3]) -> Option<CallID> {
     match call_id {
         0 => Some(CallID::ScUnknown),
-        1 => Some(CallID::ScMmap { grow: args[0] }),
-        2 => Some(CallID::ScWritev { fd: args[0], iov: args[1], iovcnt: args[2] }),
+        1 => Some(CallID::ScMmap { grow: args[0] as u32 }),
+        2 => Some(CallID::ScWritev { fd: args[0] as i32, iov: args[1] as i32, iovcnt: args[2] as u32 }),
         3 => Some(CallID::ScThreadSpawn),
         4 => Some(CallID::ScFutex),
         5 => Some(CallID::ScExit),
@@ -95,11 +95,11 @@ pub fn create_call_id(call_id: i32, args: [i32;3]) -> Option<CallID> {
 
 
 /* Convert Rust Enum CallID to instrumentation-level call_id */
-pub fn get_instrumentation_call_id(call_id: &CallID) -> (i32, [i32;3]) {
+pub fn get_instrumentation_call_id(call_id: &CallID) -> (u32, [i64;3]) {
     match call_id {
         CallID::ScUnknown => (0, [0, 0, 0]),
-        CallID::ScMmap { grow } => (1, [*grow, 0, 0]),
-        CallID::ScWritev { fd, iov, iovcnt } => (2, [*fd, *iov, *iovcnt]),
+        CallID::ScMmap { grow } => (1, [*grow as i64, 0, 0]),
+        CallID::ScWritev { fd, iov, iovcnt } => (2, [*fd as i64, *iov as i64, *iovcnt as i64]),
         CallID::ScThreadSpawn => (3, [0, 0, 0]),
         CallID::ScFutex => (4, [0, 0, 0]),
         CallID::ScExit => (5, [0, 0, 0]),
